+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added support for SQL Server
 - Added support for MongoDB
 - Added support for Redis
-- Added support for Elasticsearch
+- Added support for Elasticsearch and OpenSearch
 - Added support for international phone numbers
 
 ## 0.1.3 (2022-09-04)